+++ hadoop-hdfs-project/hadoop-hdfs/src/main/java/se/sics/hop/metadata/adaptor/INodeDALAdaptor.java
@@ -49,19 +49,19 @@ public class INodeDALAdaptor extends DALAdaptor<INode, HopINode> implements INod
     }
 
     @Override
-    public INode indexScanfindInodeById(long inodeId) throws StorageException {
+    public INode indexScanfindInodeById(int inodeId) throws StorageException {
         return convertDALtoHDFS(dataAccess.indexScanfindInodeById(inodeId));
     }
 
     @Override
-    public List<INode> indexScanFindInodesByParentId(long parentId) throws StorageException {
+    public List<INode> indexScanFindInodesByParentId(int parentId) throws StorageException {
         List<INode> list = (List) convertDALtoHDFS(dataAccess.indexScanFindInodesByParentId(parentId));
         Collections.sort(list, INode.Order.ByName);
         return list;
     }
 
     @Override
-    public INode pkLookUpFindInodeByNameAndParentId(String name, long parentId) throws StorageException {
+    public INode pkLookUpFindInodeByNameAndParentId(String name, int parentId) throws StorageException {
         return convertDALtoHDFS(dataAccess.pkLookUpFindInodeByNameAndParentId(name, parentId));
     }
