+++ hadoop-hdfs-project/hadoop-hdfs/src/main/java/se/sics/hop/metadata/context/INodeContext.java
@@ -27,12 +27,12 @@ public class INodeContext extends EntityContext<INode> {
   /**
    * Mappings
    */
-  protected Map<Long, INode> inodesIdIndex = new HashMap<Long, INode>();
+  protected Map<Integer, INode> inodesIdIndex = new HashMap<Integer, INode>();
   protected Map<String, INode> inodesNameParentIndex = new HashMap<String, INode>();
-  protected Map<Long, List<INode>> inodesParentIndex = new HashMap<Long, List<INode>>();
-  protected Map<Long, INode> newInodes = new HashMap<Long, INode>();
-  protected Map<Long, INode> modifiedInodes = new HashMap<Long, INode>();
-  protected Map<Long, INode> removedInodes = new HashMap<Long, INode>();
+  protected Map<Integer, List<INode>> inodesParentIndex = new HashMap<Integer, List<INode>>();
+  protected Map<Integer, INode> newInodes = new HashMap<Integer, INode>();
+  protected Map<Integer, INode> modifiedInodes = new HashMap<Integer, INode>();
+  protected Map<Integer, INode> removedInodes = new HashMap<Integer, INode>();
   protected INodeDataAccess<INode> dataAccess;
 
   public INodeContext(INodeDataAccess<INode>  dataAccess) {
@@ -43,8 +43,8 @@ public class INodeContext extends EntityContext<INode> {
   public void add(INode inode) throws PersistanceException {
     if (removedInodes.containsKey(inode.getId())) {
       log("added-removed-inode", CacheHitState.NA,
-              new String[]{"id", Long.toString(inode.getId()), "name", inode.getLocalName(),
-        "pid", Long.toString(inode.getParentId())});
+              new String[]{"id", Integer.toString(inode.getId()), "name", inode.getLocalName(),
+        "pid", Integer.toString(inode.getParentId())});
       removedInodes.remove(inode.getId());
       update(inode);
     } else {
@@ -52,8 +52,8 @@ public class INodeContext extends EntityContext<INode> {
       inodesNameParentIndex.put(inode.nameParentKey(), inode);
       newInodes.put(inode.getId(), inode);
       log("added-inode", CacheHitState.NA,
-              new String[]{"id", Long.toString(inode.getId()), "name", inode.getLocalName(),
-        "pid", Long.toString(inode.getParentId())});
+              new String[]{"id", Integer.toString(inode.getId()), "name", inode.getLocalName(),
+        "pid", Integer.toString(inode.getParentId())});
     }
   }
 
@@ -80,17 +80,17 @@ public class INodeContext extends EntityContext<INode> {
     INode result = null;
     switch (iFinder) {
       case ByINodeID:
-        long inodeId = (Long) params[0];
+        Integer inodeId = (Integer) params[0];
         if (removedInodes.containsKey(inodeId)) {
-          log("find-inode-by-pk-removed", CacheHitState.HIT, new String[]{"id", Long.toString(inodeId)});
+          log("find-inode-by-pk-removed", CacheHitState.HIT, new String[]{"id", Integer.toString(inodeId)});
           result = null;
         } else if (inodesIdIndex.containsKey(inodeId)) {
-          log("find-inode-by-pk", CacheHitState.HIT, new String[]{"id", Long.toString(inodeId)});
+          log("find-inode-by-pk", CacheHitState.HIT, new String[]{"id", Integer.toString(inodeId)});
           result = inodesIdIndex.get(inodeId);
         } else if (isRemoved(inodeId)) {
           return result;
         } else {
-          log("find-inode-by-pk", CacheHitState.LOSS, new String[]{"id", Long.toString(inodeId)});
+          log("find-inode-by-pk", CacheHitState.LOSS, new String[]{"id", Integer.toString(inodeId)});
           aboutToAccessStorage();
           result = dataAccess.indexScanfindInodeById(inodeId);
           inodesIdIndex.put(inodeId, result);
@@ -101,15 +101,15 @@ public class INodeContext extends EntityContext<INode> {
         break;
       case ByPK_NameAndParentId:
         String name   = (String)  params[0];
-        long parentId = (Long)    params[1];
+        Integer parentId = (Integer)    params[1];
         String key = parentId + name;
         if (inodesNameParentIndex.containsKey(key)) {
           log("find-inode-by-name-parentid", CacheHitState.HIT,
-                  new String[]{"name", name, "pid", Long.toString(parentId)});
+                  new String[]{"name", name, "pid", Integer.toString(parentId)});
           result = inodesNameParentIndex.get(key);
         } else if (newInodes.containsKey(parentId)) {
           log("find-inode-by-name-new-parentid", CacheHitState.HIT,
-                  new String[]{"name", name, "pid", Long.toString(parentId)});
+                  new String[]{"name", name, "pid", Integer.toString(parentId)});
           result = null;
         } else if (isRemoved(parentId, name)) {
           return result; // return null; the node was remove. 
@@ -119,13 +119,13 @@ public class INodeContext extends EntityContext<INode> {
           if (result != null) {
             if (removedInodes.containsKey(result.getId())) {
               log("find-inode-by-name-parentid-removed", CacheHitState.LOSS,
-                      new String[]{"name", name, "pid", Long.toString(parentId)});
+                      new String[]{"name", name, "pid", Integer.toString(parentId)});
               return null;
             }
             inodesIdIndex.put(result.getId(), result);
           }
           inodesNameParentIndex.put(key, result);
-          log("find-inode-by-name-parentid", CacheHitState.LOSS, new String[]{"name", name, "pid", Long.toString(parentId)});
+          log("find-inode-by-name-parentid", CacheHitState.LOSS, new String[]{"name", name, "pid", Integer.toString(parentId)});
         }
         break;
     }
@@ -139,12 +139,12 @@ public class INodeContext extends EntityContext<INode> {
     List<INode> result = null;
     switch (iFinder) {
       case ParentId:
-        long parentId = (Long) params[0];
+        Integer parentId = (Integer) params[0];
         if (inodesParentIndex.containsKey(parentId)) {
-          log("find-inodes-by-parentid", CacheHitState.HIT, new String[]{"pid", Long.toString(parentId)});
+          log("find-inodes-by-parentid", CacheHitState.HIT, new String[]{"pid", Integer.toString(parentId)});
           result = inodesParentIndex.get(parentId);
         } else {
-          log("find-inodes-by-parentid", CacheHitState.LOSS, new String[]{"pid", Long.toString(parentId)});
+          log("find-inodes-by-parentid", CacheHitState.LOSS, new String[]{"pid", Integer.toString(parentId)});
           aboutToAccessStorage();
           result = syncInodeInstances(dataAccess.indexScanFindInodesByParentId(parentId));
           Collections.sort(result, INode.Order.ByName);
@@ -191,7 +191,7 @@ public class INodeContext extends EntityContext<INode> {
     newInodes.remove(inode.getId());
     modifiedInodes.remove(inode.getId());
     removedInodes.put(inode.getId(), inode);
-    log("removed-inode", CacheHitState.NA, new String[]{"id", Long.toString(inode.getId()), "name", inode.getLocalName()});
+    log("removed-inode", CacheHitState.NA, new String[]{"id", Integer.toString(inode.getId()), "name", inode.getLocalName()});
   }
 
   @Override
@@ -209,7 +209,7 @@ public class INodeContext extends EntityContext<INode> {
     inodesIdIndex.put(inode.getId(), inode);
     inodesNameParentIndex.put(inode.nameParentKey(), inode);
     modifiedInodes.put(inode.getId(), inode);
-    log("updated-inode", CacheHitState.NA, new String[]{"id", Long.toString(inode.getId()), "name", inode.getLocalName()});
+    log("updated-inode", CacheHitState.NA, new String[]{"id", Integer.toString(inode.getId()), "name", inode.getLocalName()});
   }
 
   private List<INode> syncInodeInstances(List<INode> newInodes) {
@@ -243,7 +243,7 @@ public class INodeContext extends EntityContext<INode> {
     return finalList;
   }
 
-  private boolean isRemoved(final long parent_id, final String name) {
+  private boolean isRemoved(final Integer parent_id, final String name) {
     for (INode inode : removedInodes.values()) {
       if (inode.getParentId() == parent_id
               && inode.getLocalName().equals(name)) {
@@ -254,7 +254,7 @@ public class INodeContext extends EntityContext<INode> {
     return false;
   }
 
-  private boolean isRemoved(final long id) {
+  private boolean isRemoved(final Integer id) {
     for (INode inode : removedInodes.values()) {
       if (inode.getId() == id) {
         return true;
@@ -278,7 +278,7 @@ public class INodeContext extends EntityContext<INode> {
         //delete the previous row from db
         INode inode = (INode) params[0];
         removedInodes.put(inode.getId(),inode);
-        log("snapshot-maintenance-removed-inode", CacheHitState.NA, new String[]{"id", Long.toString(inode.getId()), "name", inode.getLocalName(), "pid", Long.toString(inode.getParentId()) });
+        log("snapshot-maintenance-removed-inode", CacheHitState.NA, new String[]{"id", Integer.toString(inode.getId()), "name", inode.getLocalName(), "pid", Integer.toString(inode.getParentId()) });
         break;
     }
   }
